requirements.txt
requests
python-dotenv
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
gspread
cryptography
flask
gunicorn
google-cloud-secret-manager
google-cloud-firestore
